requirements.txt
PyQt5==5.15.10
Pillow==10.1.0
python-dotenv==1.0.0
pyinstaller==6.1.0
